backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
google-generativeai
chromadb
pypdf
python-docx
openpyxl
python-pptx
langchain
PyJWT